Add HasOneThrough relation to QuickCount model
A HasOneThrough Eloquent relation has been added to the QuickCount model to establish a direct relation with the 'Caleg' model through the 'Partai' model. This replaces the previous BelongsToMany relation. The change has been reflected in the QuickCountResource file by updating the path for the 'calegPartai' attribute. This is done to optimize the database querying and improve readability of the code.