Add Dutch (NL) and Polish (PL) translations
In this commit, translations in Dutch (NL) and Polish (PL) were added to multiple pages. This is necessary to extend the localization support and to provide a more user-friendly experience for Dutch and Polish speaking users. These translations cover various pages such as authentication (login, register, and profile adjustment), tenancy profile edit, and record management (creation, editing, viewing, and listing of records).